separate SuiteRunner class
